Focus Zone: fixing empty tabbable element
There's code in Focus Zone written in a way such that once a default tabbable element is set, it can never be reset, no matter if the user of the focus zone passes a `undefined` defaultTabbableElement, or not pass it at all. Let's tweak that a little so that not-passing\resetting the defaultTabbableElement has the same effect as never having it set in the first place.

Testing:

* Added action in the FocusZoneTest component to test the effect of setting any of the nine buttons as the default tabbable element plus another to clear it and checked that clearing the default tabbable makes the focus zone behave as expected (as if one was never set)
* Checked this sequence: set -> clear -> set -> clear