Fix usage of conditional hooks (#841)
Fix usage of conditional hooks

There are various places in code where we do something like the following, which probably intends to read as grab the `componentRef` from `userProps` if present, otherwise make a new one:

`const { componentRef = React.useRef(null), ... rest } = userProps`

However, this is problematic if during one render pass a given component does not receive `componentRef` (so we make a ref) and in the next render pass it does (so we never try to use that ref we made earlier). The reason this is specifically problematic is that the `React.useRef(null)` **will not be called** if we were successful in extracting `componentRef` from `props`.

The implication here is that we invalidate a very important [rule of hooks](https://fb.me/rules-of-hooks/) -- that they must always be called in the same order. This will in turn lead to unexpected error messages in downstream\client code that uses Fluent components like the below:

`Warning: React has detected a change in the order of Hooks called by Button. This will lead to bugs and errors if not fixed. For more information, read the Rules of Hooks: https://fb.me/rules-of-hooks

   Previous render            Next render
   ------------------------------------------------------
1. useMemo                    useMemo
2. useRef                     useCallback
   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^`

To fix, let's just make sure we always call `React.useRef(null)` -- we can still toss it aside if not wanted by slightly tweaking how the code is structured.

Testing:

* Sanity tested button, checkbox, and radio button components in the Win32 fluent tester
* I didn't attempt to repro the problem in the Fluent tester, but did validate the fix in the product where I originally discovered this problem.