Declare tslib dependency (#1085)
Fixes #1025

### Platforms Impacted
- [ ] iOS
- [ ] macOS
- [ ] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

Declare tslib dependency where tslib is used and we have importHelpers in tsconfig.ts in our framework package. Version of tslib is ^1.13.0 as that's the latest release that's supported due to our current typescript version (we need 3.9 or above to bump to 2.0 or later)

Also took the opportunity to use importHelpers in other packages that could use it, specifically for spreadArray so that is saves on bundle size. But if that change isn't desired, I can undo the last two commits and just fix the existing importHeaders

### Verification

Change builds, verifying that the bundle works properly on win32 and windows.

| Before                                       | After                                      |
|----------------------------------------------|--------------------------------------------|
| Screenshot or description before this change | Screenshot or description with this change |

### Pull request checklist

This PR has considered (when applicable):
- [ ] Automated Tests
- [ ] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts