Add Scrollview and FocusZone slots to ContextualMenu (#1409)
### Platforms Impacted
- [ ] iOS
- [x] macOS
- [x] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

This is the first of a few changes that I'm breaking https://github.com/microsoft/fluentui-react-native/pull/1400 apart into.

This change refactors the subviews "ScrollView" and "FocusZone" into slots. Doing so allows us to set up and reference their state in `usePrepareProps`. This would be useful if I want to say, set a componentRef on FocusZone that I want to access later on. 

This change also makes scrollView a required slot, and adds it to the macOS contextualMenu. Turns out that's fine, things still behave as they did before. The FocusZone slot is made optional, as it is only needed on macOS (where we can't move focus with arrow keys by default). 

### Verification

This change should largely be a no-op, and CI's E2E tests should catch regressions (as they did in #1400).

### Pull request checklist

This PR has considered (when applicable):
- [ ] Automated Tests
- [ ] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts