Implement gap token on MenuList (#1680)
### Platforms Impacted
- [ ] iOS
- [ ] macOS
- [ ] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

MenuList has a gap between items on web. Implementing that token on MenuList.

Stack would be handy to use in this situation, since it has a gap prop. Unfortunately the Stack component that's in FURN is not working on the v1 framework (no idea why) and I'd like to avoid using v0 components where possible. It would make sense to transition to using a Stack component for this in the future but for now, the gap logic was fairly easy to implement so I just added a hidden MenuStack component.

### Verification
gap set to something like 6:

![image](https://user-images.githubusercontent.com/4602628/169629041-5d68a65b-14f2-40db-9524-79f67c7ccd66.png)

### Pull request checklist

This PR has considered (when applicable):
- [ ] Automated Tests
- [ ] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts