Datepicker - add ability to update the reference start/end dates (#1687)
### Platforms Impacted
- [x] iOS
- [ ] macOS
- [ ] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

Ask from a client team to be able to set the reference start/end dates (range of possible dates the calendar allows, not to be confused with "date range" which refers to the dates a user can select). 

The default reference start date is Jan 1, 3 years before the current year. The default reference end date is 10 years after the default reference start date. Ex. For May 23, 2022, the default range is Jan 1, 2019 - Dec 21, 2029.

This change allows the default Calendar Configuration to be overridden - which includes the reference start and end dates, as well as firstWeekday (usually begins on Sunday or Monday). Note that overriding the default calendar configuration affects **all** FURN datepickers. Additional work will be required if we want to support different calendar configurations for different datepickers.

### Verification

Added a button at the bottom of the FluentTester DatePicker test page that updates the default reference start/end dates to be from 1999 to 2000.

| Before | After |
|----------------------------------------------|--------------------------------------------|
| ![image](https://user-images.githubusercontent.com/78454019/169923134-4bc8469e-2b98-48dc-9595-42ce730fd8d3.png) | ![image](https://user-images.githubusercontent.com/78454019/169923462-d6d96d36-0931-4186-8230-0a07d1e901e2.png) | ![image](https://user-images.githubusercontent.com/78454019/169923134-4bc8469e-2b98-48dc-9595-42ce730fd8d3.png)
 

| Before (earliest possible start date you could scroll to) | After (earliest possible start date you could scroll to) |
|----------------------------------------------|--------------------------------------------|
| ![image](https://user-images.githubusercontent.com/78454019/169923204-43860ae9-bf22-46ed-82aa-4dd75d33a5b1.png) | ![image](https://user-images.githubusercontent.com/78454019/169923219-51e41698-0804-4d80-9602-6134cc630aaa.png) |

### Pull request checklist

This PR has considered (when applicable):
- [ ] Automated Tests
- [ ] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts


Notes:
- Not sure if using a button in the FluentTester datepicker page to change the default reference start end dates is best. Maybe a toggle would be better?
- This allows users to set the default "CalendarConfiguration", but this PR only exposes referenceStartDate and referenceEndDate. Do we also want to expose firstWeekday?
- With this change it is easier to get into a state where the selected date is outside the reference start/end dates, ex if the default start date is already out of the range, and you try to select the default date. There is an assert in datetimepickerviewdatasource (line 308, assert(dateIndex >= 0 && dateIndex < numberOfDates, "index out of bound")) that gets hit when this happens. Do we want to do anything about this?