Remove menu popover from being tab stop (#1781)
### Platforms Impacted
- [ ] iOS
- [ ] macOS
- [x] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

Bug was that the menu popover was a tab stop when tabbing when the menu was opened via mouse. This was because it was listed as tabble when the Menu is invoked via mouse.

Copied this fix from ContextualMenu. Focusable is set to false on the container as soon as focus is shifted away from it, removing it from the tab loop.

### Verification

Tested via tester.

### Pull request checklist

This PR has considered (when applicable):
- [ ] Automated Tests
- [ ] Documentation and examples
- [x] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts