[Button] Prefer user key events over those handled by Fluent (#1931)
* [Button] Prefer user key events over those handled by Fluent

Currently, on Windows, the `Button` component handles keyUp events
(keyDown for macOS). What that means is someone can't have a keyUp (or
keyDown) handler of their own to do "special" things on keypresses on
buttons (e.g. perhaps implement a split button).

Previously (prior to PR #1373), `Button` always used keyUp, so button
users could at least use keyDown to do things, but now depending on the
target platform, a button user would have to handle the slot NOT used by
Fluent.

I investigated the possibility of creating a composite handler (such
that for user requested keys we always pass back to the user, and if not
specified, then we pass back the handling to Fluent). It looks like it
will need substantial refactoring because the event types for key
handlers in useKeyProps does not match that exposed by Win32. Further,
the approach to specifying valid keys (the name of the prop, as well the
specifics like modifiers) is different for the Mac.

Therefore, in this change, I am instead suggesting the *other* approach
I previously discussed with Saad -- if a button user tries to override a
particular key handler, the user is also responsible for providing the
default behaviors (like Enter\Space to invoke the button event).

* Expose a flag to indicate platform behavior w.r.t. keyUp or keyDown

* Change files

* update docs