Restore ability to create unfocusable buttons
Previously (with the old deprecated buttons) it was possible to create buttons that aren't disabled but still are unfocusable. (this may sound like a bizarre use case, but is handy for creating things like split buttons where the individual component buttons are not focusable).

We seem to have lost this ability with the new Button -- which will override focusable even if passed by the component user. Let's not -- if that is specified, we respect it (and allow people to create un-focusable enabled buttons, or focusable disabled buttons as before).

Testing:

* Manually, in the test app.