Shadow: don't render the extra view if there isn't a shadow (#2001)
### Platforms Impacted
- [x] iOS
- [x] macOS
- [ ] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

Follow up to a conversation in #1897 

Just return the child if there is no shadow (i.e. if ShadowToken is undefined) - don't add an extra view, or add unnecessary shadow props. Also updated Notification to pass in undefined as well.

Passing in undefined to not render is a Shadow isn't ideal, but not sure what a better solution would be. The other idea we had was to add a shadow0 token to the token pipeline, which would work too, it just seems a little strange because I don't think we'd ever actually use any of the values. Still though, if passing in undefined is confusing we could add support so that if either undefined or shadow0 are passed in, we just return the child.

### Verification

No visible changes, just verified with the snapshot tests. Added a test for a FAB that doesn't have a shadow, first commit contains snapshot of what this snapshot used to be before the change, second commit contains the change + what the snapshot looks like after the change - no longer has the extra view or the extra shadow props.

### Pull request checklist

This PR has considered (when applicable):
- [ ] Automated Tests
- [ ] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts