Shadow: Restrict child to only take type View (#2096)
### Platforms Impacted
- [x] iOS
- [x] macOS
- [ ] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

This is a follow up to investigation that happened here: https://github.com/microsoft/react-native-macos/pull/1402

After fixing a bug related to shadow consistency in react-native-macOS, one of the tests in react-native-macOS started failing. The issue seemed to be with the test itself (and its still unclear why the test only started failing after the bug fix) - the test was setting shadow props (ex. shadowColor, shadowOpacity, shadowOffset, shadowRadius) on a Text element, when [Text actually has its own specific shadow props](https://reactnative.dev/docs/text-style-props#textshadowcolor) - i.e. textShadowColor, textShadowOffset, textShadowRadius (interestingly textShadowOpacity is not a thing)

I considered updating the Shadow component to also work for Text, but decided against it because Shadows on Text doesn't seem to be an actual use case - usually it's a Shadow on a View that has text on it.

This PR
- updates the jest tests that were doing the same thing as the test in react-native-macOS, the Text should be wrapped in a view and then the Shadow is wrapped around that
- updates the documentation to reflect this change

### Verification

I tried the following code that was in the Shadow jest tests in FluentTester:

```javascript
<Shadow shadowToken={theme.shadows.shadow64}>
    <Text>
        test
    </Text>
</Shadow>
```
and this actually caused a red box to appear (same error as in the react-native-macOS test that started failing)

<img width="335" alt="Screen Shot 2022-09-07 at 12 02 33 PM" src="https://user-images.githubusercontent.com/78454019/188957094-140956b4-80c0-479e-95eb-814f19d29277.png">

Changing the code to this stopped the red box from showing up.
```javascript
<Shadow shadowToken={theme.shadows.shadow64}>
    <View>
        <Text>
            test
        </Text>
    </View>
</Shadow>
```

I would have expected the jest tests to fail on any code that causes a red box, if anyone has any insight on why the jest tests did not catch this please let me know (best guess is because the red box error is coming from native code, and jest tests don't catch those).

No visible changes in the tester.

### Pull request checklist

This PR has considered (when applicable):
- [x] Automated Tests
- [x] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts