Change strategy in finding search root for queries
Previously, we designated a root view to start our searches from. For most
elements, this works, but in cases such as ContextualMenu on win32, those
children can't be found under our root, meaning we have to search the whole
tree. There are similar cases with UWP.

This commit changes how we find our search root by separating logic per
platform. UWP and Win32 e2etests have different ways in which they can
find the top window to start searches from. This way, there does not need
to be an edge case for ContextualMenuItems, and this way, we don't need
chained awaits.

Currently, strategies for finding a starting root for searches still need
to be implemented for iOS, macOS, and Android.