Use selector chaining to improve element query performance (#2226)
* Add selector chaining in `By` method using root

* Add await keyword to `By` func calls

* Change files

* Fix bug in searching for elements outside rootView

In some instances, elements we're looking for, such as some menu
items, aren't children of the rootView we designate. This commit
adds an if statement to check for this.

This commit also fixes missing await keywords in NavigateAppPage and
NativeTestingPageObject.

* Add fallback for UWP failing pipeline

* Add environment var for platform-specific querying

* Change scrollToComponentButton to use environment var

* Use query chaining for ios tests

* Fix iOS tests failing

* Change files

* Change condition which makes root view accessible

* Integrate experimental link tests

* Fix e2etests failing in win ci

* Rename _rootView -> rootView

* Remove chained awaits from Query calls

* Revert "Remove chained awaits from Query calls"

This reverts commit 188c7f9cf41f7a7fdd5986adc28801b3e3121369.

* Change strategy in finding search root for queries

Previously, we designated a root view to start our searches from. For most
elements, this works, but in cases such as ContextualMenu on win32, those
children can't be found under our root, meaning we have to search the whole
tree. There are similar cases with UWP.

This commit changes how we find our search root by separating logic per
platform. UWP and Win32 e2etests have different ways in which they can
find the top window to start searches from. This way, there does not need
to be an edge case for ContextualMenuItems, and this way, we don't need
chained awaits.

Currently, strategies for finding a starting root for searches still need
to be implemented for iOS, macOS, and Android.

* Remove leftovers from testing

Removed leftover code from testing. Also adjusted UWP root
search strategy to be more specific.

* Revert "Remove leftovers from testing"

This reverts commit 04e6b45fb9f052258711f69a3756b0874c69a0fc.

* Revert "Change strategy in finding search root for queries"

This reverts commit 4269a0f28774890229d73289ac701ef59864c4af.

* Fix nits

Co-authored-by: Lawrence Win <winlarry@microsoft.com>