Trying to avoid Win32 E2E failure
The problem seems to be that adding another button adjusts layout enough that BasePage.scrollToTestElement fails to do the right thing (tm). Specifically, there is logic in there that hits PgDown until the element in question is visible, after which it does another PgDown to "ensure the whole component is visible". Unfortunately that is what scrolls the expected element OUT of view, at which point we then start getting failures around "The primary UI element for testing did not display correctly." (well, it did, it's just out of viewport).

I imagine the fix is along the lines of not doing the extra PgDown if the element is fully visible (and if it is partially visible, doing a smaller scroll, not PgDown, which is in all probability too much). But because of the potential scope of this change, I elect to not address that in this pull request. Instead, I will change the accessibilityRole on another, existing button. The automated tests we have appear to only check the accessibility label on this control.