[Switch] Disable state a11y fix (#2494)
### Platforms Impacted
- [x] iOS
- [ ] macOS
- [ ] win32 (Office)
- [ ] windows
- [x] android

### Description of changes
This PR targets to fix the a11y issue where keyboard navigation also goes into disabled state switch which shouldn't happen. 
This fixes it for mobile platform - Android and iOS. 

This was already working on win32, hence verified that it doesn't break it and working as before. 

### Verification

Visually verified on Android, tested on win32 as well, it's working as before. 


| Before                                       | After                                      |
|----------------------------------------------|--------------------------------------------|
| [Screen Recording (2).webm](https://user-images.githubusercontent.com/30728574/211253264-09dd856e-998c-4b9e-b2f1-a89a3ebfa623.webm) | [Screen Recording (1).webm](https://user-images.githubusercontent.com/30728574/211253237-e2d08847-7064-4995-bf0c-df798f4ef4d5.webm) |

### Pull request checklist

This PR has considered (when applicable):
- [ ] Automated Tests
- [x] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts