Rewrite animated indicator to work better with RTL
This commit simplifies the logic of the animated indicator. Rather than
positioning the indicator using a calculated `start` value, we position
it using absolute layout values within the tablist. Now, the
`useTabAnimation` hook does all the math to figure out the top and left
offsets of the indicator, and the styling hook for the animatedIndicator
simply uses them to position the `top` and `left` layout props.

The container slot of the animated indicator is also removed.